+++ tests/test_threadpoolctl.py
@@ -1,13 +1,14 @@
 from __future__ import annotations
 
+import ctypes
 import json
 import os
 import pytest
 import re
 import subprocess
-import sys
-import ctypes
 import shutil
+import sys
+from threading import Thread
 
 import threadpoolctl
 from threadpoolctl import threadpool_limits, threadpool_info
@@ -1064,3 +1065,38 @@ def test_conda_blas_detection_after_import(module):
         if each["internal_api"] != "flexiblas"
     ]
     assert set(blas_names_from_threadpoolctl).issubset(blas_names_from_conda)
+
+
+def test_controller_parallelism_no_deadlocks():
+    """Creating a controller in parallel to itself does not cause deadlocks.
+
+    Non-regression test for https://github.com/joblib/threadpoolctl/issues/239
+
+    Lacking the fixes from PR #243, this deadlocks on Conda environments, at
+    least, but possibly not on PyPI with Python from a Linux distro.
+    """
+    if sys.platform != "linux" or not hasattr(ctypes.PyDLL(None), "backtrace"):
+        pytest.skip("Testing glibc on Linux")
+
+    # Internally, backtrace() calls dl_iterate_phdr which can result in
+    # deadlocks if threadpoolctl is also using dl_iterate_phdr.
+    backtrace_gil = ctypes.PyDLL(None).backtrace
+    backtrace_gil.argtypes = [ctypes.c_void_p, ctypes.c_int]
+    backtrace_nogil = ctypes.CDLL(None).backtrace
+    backtrace_nogil.argtypes = [ctypes.c_void_p, ctypes.c_int]
+
+    def create_controllers():
+        buf = (ctypes.c_void_p * 20)()
+        for _ in range(100):
+            limiter = threadpool_limits()
+            backtrace_gil(buf, 20)
+            backtrace_nogil(buf, 20)
+
+    threads = []
+    for _ in range(os.cpu_count() * 4):
+        t = Thread(target=create_controllers)
+        threads.append(t)
+        t.start()
+
+    for t in threads:
+        t.join()