+++ CHANGES.md
@@ -41,6 +41,13 @@
 - Fix OpenBLAS detection for conda package on Windows
   https://github.com/joblib/threadpoolctl/pull/240
 
+- Fixed a deadlock on Linux when using threadpoolctl from multiple threads.
+  https://github.com/joblib/threadpoolctl/pull/243
+
+  - Start using Python 3.14's built-in support for listing shared libraries.
+
+  - On Linux, start using /proc/self/maps for listing shared libraries.
+
 3.6.0 (2025-03-13)
 ==================